WebApp/requirements.txt
azure-core==1.34.0
azure-identity==1.23.0
azure-storage-blob==12.25.1
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.2
dataclass-wizard==0.35.0
Flask==3.1.1
greenlet==3.2.2
htmlmin==0.1.12
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msal==1.32.3
msal-extensions==1.3.1
pycparser==2.22
PyJWT==2.10.1
pyodbc==5.2.0
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
